requirements.txt
customtkinter
requests
geopy
pytz
timezonefinder
Pillow